Add unit tests

+++ src/components/Cart/Cart.tsx
@@ -6,16 +6,16 @@ import { Pokemon } from '../../types/Common.types';
 import { useCallback } from 'react';
 import { AppState } from '../../store/store.types';
 
-const Cart = ({ open, onClose, onRemove }: CartProps) => {
+const Cart = ({ open = false, onClose, onRemove }: CartProps) => {
     const items = useSelector(({ cart }: AppState) => cart);
 
     const handleClose = useCallback(() => {
-        onClose();
-    }, []);
+      onClose && onClose();
+    }, [onClose]);
 
     const handleRemove = useCallback((pokemon: Pokemon) => {
-        onRemove(pokemon);
-    }, []);
+      onRemove && onRemove(pokemon);
+    }, [onRemove]);
 
     let subtotal = 0;
     let totalItems = 0;
@@ -32,7 +32,7 @@ const Cart = ({ open, onClose, onRemove }: CartProps) => {
 
         const formatedPrice = numberFormat.format(item.price);
         return (
-            <li key={item.name} className={styles.cartItem}>
+            <li key={item.name} className={styles.cartItem} data-testid={item.name}>
                 <div className={styles.cartItemImageWrapper}>
                     <img
                         alt={item.name}
@@ -43,7 +43,13 @@ const Cart = ({ open, onClose, onRemove }: CartProps) => {
                 <div className={styles.cartItemDetails}>
                     <p>{item.name}</p>
                     <p>{`x${item.amount} ${formatedPrice}`}</p>
-                    <button type="button" onClick={() => handleRemove(item)}>Remover</button>
+                    <button
+                      type="button"
+                      onClick={() => handleRemove(item)}
+                      data-testid={`${item.name}-button`}
+                    >
+                      Remover
+                    </button>
                 </div>
             </li>
         )
@@ -51,16 +57,19 @@ const Cart = ({ open, onClose, onRemove }: CartProps) => {
 
     const subtotalFormated = numberFormat.format(subtotal);
     return(
-        <div className={`${styles.cartWrapper} ${open ? `${styles.open}` : ''}`}>
-            <button type="button" onClick={handleClose}>Fechar</button>
-            <div className={styles.scrollableContent}>
-                <ul>
-                    {map}
-                </ul>
-            </div>
-            <p>{`Items: ${totalItems}`}</p>
-            <p>{`Subtotal: ${subtotalFormated}`}</p>
+      <div
+        className={`${styles.cartWrapper} ${open ? `${styles.open}` : ''}`}
+        data-testid="cart"
+      >
+        <button type="button" onClick={handleClose} data-testid="cart-button-close">Fechar</button>
+        <div className={styles.scrollableContent}>
+            <ul>
+                {map}
+            </ul>
         </div>
+        <p>{`Items: ${totalItems}`}</p>
+        <p>{`Subtotal: ${subtotalFormated}`}</p>
+      </div>
     )
 }
